% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depguard-package.R
\name{print.depguard_snapshot}
\alias{print.depguard_snapshot}
\title{Print a depguard snapshot}
\usage{
\method{print}{depguard_snapshot}(x, ...)
}
\arguments{
\item{x}{A \code{depguard_snapshot} object.}

\item{...}{Ignored.}
}
\value{
The \code{x} object, returned invisibly.
}
\description{
Print a depguard snapshot
}
