% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{dep_manifest_read}
\alias{dep_manifest_read}
\title{Read an existing dependency manifest}
\usage{
dep_manifest_read(path = default_manifest_path())
}
\arguments{
\item{path}{File path to the manifest. Defaults to
\code{.depguard_manifest.rds} in the current working directory.}
}
\value{
A named character vector of package requirements, or \code{NULL}
(invisibly) with a message if no manifest is found.
}
\description{
Read an existing dependency manifest
}
